+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.13.0] - 2026-02-13
+### Changed
+- Updated all compiled protos for compatibility with Injective core v1.18.0 and Indexer v1.18.3
+- Includes new proto definitions for the Chainlink Data Streams oracle
+
 ## [1.12.0] - 2025-11-10
 ### Changed
 - Updated all compiled protos for compatibility with Injective core v1.17.0 and Indexer v1.17.16